rf-17: file renaming
